Convert jacoco configurations to buildsrc
