+++ buildSrc/src/main/kotlin/plugin/MultiplatformConfigurationPlugin.kt
@@ -12,6 +12,7 @@ class MultiplatformConfigurationPlugin : Plugin<Project> {
     override fun apply(project: Project) {
         configureMpp(project)
         project.configureMppTest()
+        project.configureJacocoMultiplatform()
         project.configureMavenMultiplatform()
         project.configureBintray()
     }
